[otbn] Split snippet generators into a separate directory
No functional change except for a re-organisation, but we're going to
add more generators over time, so we are going to need them to be in
separate files to avoid a mega-module.
